+++ Lilu/Sources/kern_patcher.cpp
@@ -156,16 +156,22 @@ size_t KernelPatcher::loadKinfo(KernelPatcher::KextInfo *info) {
 }
 #endif /* KEXTPATH_SUPPORT */
 
-void KernelPatcher::updateRunningInfo(size_t id, mach_vm_address_t slide, size_t size, bool force) {
+size_t KernelPatcher::updateRunningInfo(size_t id, mach_vm_address_t slide, size_t size, bool force) {
 	if (id >= kinfos.size()) {
 		SYSLOG("patcher", "invalid kinfo id %lu for running info update", id);
-		return;
+		return size;
 	}
 
 	if (kinfos[id]->getRunningAddresses(slide, size, force) != KERN_SUCCESS) {
 		SYSLOG("patcher", "failed to retrieve running info");
 		code = Error::KernRunningInitFailure;
 	}
+
+	// In 11.0 khandler's size only contains __TEXT. We need to update this.
+	uint8_t *hdr = nullptr;
+	size_t nsize = 0;
+	kinfos[id]->getRunningPosition(hdr, nsize);
+	return nsize > size ? nsize: size;
 }
 
 bool KernelPatcher::compatibleKernel(uint32_t min, uint32_t max) {
@@ -224,7 +230,7 @@ void KernelPatcher::setupKextListening() {
 	// We have already done this
 	if (that) return;
 
-	loadedKextSummaries = reinterpret_cast<OSKextLoadedKextSummaryHeader **>(solveSymbol(KernelID, "_gLoadedKextSummaries"));
+	loadedKextSummaries = reinterpret_cast<OSKextLoadedKextSummaryHeaderAny **>(solveSymbol(KernelID, "_gLoadedKextSummaries"));
 
 	if (loadedKextSummaries) {
 		DBGLOG("patcher", "_gLoadedKextSummaries address %p", loadedKextSummaries);
@@ -372,7 +378,7 @@ mach_vm_address_t KernelPatcher::routeFunctionShort(mach_vm_address_t from, mach
 	return routeFunctionInternal(from, to, buildWrapper, kernelRoute, revertible, JumpType::Short);
 }
 
-mach_vm_address_t KernelPatcher::routeFunctionInternal(mach_vm_address_t from, mach_vm_address_t to, bool buildWrapper, bool kernelRoute, bool revertible, JumpType jumpType) {
+mach_vm_address_t KernelPatcher::routeFunctionInternal(mach_vm_address_t from, mach_vm_address_t to, bool buildWrapper, bool kernelRoute, bool revertible, JumpType jumpType, MachInfo *info) {
 	mach_vm_address_t diff = (to - (from + SmallJump));
 	int32_t newArgument = static_cast<int32_t>(diff);
 
@@ -396,6 +402,7 @@ mach_vm_address_t KernelPatcher::routeFunctionInternal(mach_vm_address_t from, m
 	// If we already routed this function, we simply redirect the original function
 	// to the new one, and call the previous function as "original".
 	mach_vm_address_t trampoline = readChain(from);
+	mach_vm_address_t addressSlot = 0;
 	if (trampoline) {
 		// Do not perform double revert
 		revertible = false;
@@ -403,13 +410,26 @@ mach_vm_address_t KernelPatcher::routeFunctionInternal(mach_vm_address_t from, m
 		// is an unsupported configuration, as it breaks previous plugin...
 		if (!buildWrapper) trampoline = 0;
 	} else if (buildWrapper) {
-		trampoline = createTrampoline(from, absolute ? LongJump : SmallJump);
+		if (info && absolute && jumpType == JumpType::Auto) {
+			addressSlot = info->getAddressSlot();
+			DBGLOG("patcher", "using slotted jumping via " PRIKADDR, CASTKADDR(addressSlot));
+		}
+		trampoline = createTrampoline(from, absolute ? (addressSlot ? MediumJump : LongJump) : SmallJump);
 		if (!trampoline) return EINVAL;
 	}
 
 	// In case we already a trampoline to return, do not return it.
 	Patch::All *opcode = nullptr, *argument = nullptr, *disp = nullptr;
-	if (absolute) {
+	// The reason to use slots is to reduce the patch size even for absolute patches.
+	// In this case we store 6 bytes of indirect jmp with the target address itself being
+	// put in the beginning of the image, right after the Mach-O commands.
+	// This solves the problem of having short prologues followed by non-movable
+	// commands like mov rax, <vtable_address> commonly found in 11.0 (e.g. ATIController::start).
+	if (addressSlot) {
+		opcode = Patch::create<Patch::Variant::U16>(from, LongJumpPrefix);
+		argument = Patch::create<Patch::Variant::U32>(from + sizeof(LongJumpPrefix), static_cast<uint32_t>(addressSlot - (from + MediumJump)));
+		disp = Patch::create<Patch::Variant::U64>(addressSlot, to);
+	} else if (absolute) {
 		opcode = Patch::create<Patch::Variant::U16>(from, LongJumpPrefix);
 		argument = Patch::create<Patch::Variant::U32>(from + sizeof(LongJumpPrefix), 0);
 		disp = Patch::create<Patch::Variant::U64>(from + sizeof(LongJumpPrefix) + sizeof(uint32_t), to);
@@ -450,9 +470,11 @@ mach_vm_address_t KernelPatcher::routeFunctionInternal(mach_vm_address_t from, m
 				return trampoline;
 
 			SYSLOG("patcher", "failed to store patches for later removal, you are in trouble");
+#ifndef __clang_analyzer__
 			if (oidx) kpatches.erase(oidx);
 			if (aidx) kpatches.erase(aidx);
 			if (didx) kpatches.erase(didx);
+#endif
 		}
 	}
 
@@ -526,7 +548,7 @@ bool KernelPatcher::routeMultipleInternal(size_t id, RouteRequest *requests, siz
 		auto &request = requests[i];
 		if (!request.from) continue;
 		if (request.to) eraseCoverageInstPrefix(request.from, 5, LongJump);
-		auto wrapper = routeFunctionInternal(request.from, request.to, request.org, kernelRoute, true, jump);
+		auto wrapper = routeFunctionInternal(request.from, request.to, request.org, kernelRoute, true, jump, kinfos[id]);
 		if (request.org) {
 			if (wrapper) {
 				DBGLOG("patcher", "wrapped %s", request.symbol);
@@ -557,8 +579,10 @@ uint8_t KernelPatcher::tempExecutableMemory[TempExecutableMemorySize] __attribut
 
 mach_vm_address_t KernelPatcher::readChain(mach_vm_address_t from) {
 	// Note, unaligned access for simplicity
-	if (*reinterpret_cast<decltype(&LongJumpPrefix)>(from) == LongJumpPrefix)
-		return *reinterpret_cast<mach_vm_address_t *>(from + sizeof(LongJumpPrefix) + sizeof(uint32_t));
+	if (*reinterpret_cast<decltype(&LongJumpPrefix)>(from) == LongJumpPrefix) {
+		auto disp = *reinterpret_cast<int32_t *>(from + sizeof(LongJumpPrefix));
+		return *reinterpret_cast<mach_vm_address_t *>(from + sizeof(MediumJump) + disp);
+	}
 	if (*reinterpret_cast<decltype(&SmallJumpPrefix)>(from) == SmallJumpPrefix)
 		return from + SmallJump + *reinterpret_cast<int32_t *>(from + sizeof(SmallJumpPrefix));
 	return 0;
@@ -643,14 +667,15 @@ void KernelPatcher::onKextSummariesUpdated() {
 		DBGLOG("patcher", "invoked at kext loading/unloading");
 
 		if (that->activated && that->loadedKextSummaries) {
-			auto num = (*that->loadedKextSummaries)->numSummaries;
+			auto num = (*that->loadedKextSummaries)->base.numSummaries;
 			if (num > 0) {
 				if (that->waitingForAlreadyLoadedKexts) {
-					that->processAlreadyLoadedKexts((*that->loadedKextSummaries)->summaries, num);
+					that->processAlreadyLoadedKexts((*that->loadedKextSummaries), num);
 					that->waitingForAlreadyLoadedKexts = false;
 				}
 				if (that->khandlers.size() > 0) {
-					OSKextLoadedKextSummary &last = (*that->loadedKextSummaries)->summaries[num-1];
+					OSKextLoadedKextSummaryBase &last = getKernelVersion() >= KernelVersion::BigSur
+						? (*that->loadedKextSummaries)->bigSur.summaries[num-1].base : (*that->loadedKextSummaries)->legacy.summaries[num-1].base;
 					DBGLOG("patcher", "last kext is " PRIKADDR " and its name is %.*s", CASTKADDR(last.address), KMOD_MAX_NAME, last.name);
 					// We may add khandlers items inside the handler
 					for (size_t i = 0; i < that->khandlers.size(); i++) {
@@ -682,11 +707,12 @@ void KernelPatcher::onKextSummariesUpdated() {
 	}
 }
 
-void KernelPatcher::processAlreadyLoadedKexts(OSKextLoadedKextSummary *summaries, size_t num) {
+void KernelPatcher::processAlreadyLoadedKexts(OSKextLoadedKextSummaryHeaderAny *header, size_t num) {
 	DBGLOG("patcher", "processing already loaded kexts by iterating over %lu summaries", num);
 
 	for (size_t i = 0; i < num; i++) {
-		auto curr = summaries[i];
+		OSKextLoadedKextSummaryBase &curr = getKernelVersion() >= KernelVersion::BigSur
+			? header->bigSur.summaries[i].base : header->legacy.summaries[i].base;
 		for (size_t j = 0; j < khandlers.size(); j++) {
 			auto handler = khandlers[j];
 			if (handler->loaded && !strncmp(handler->id, curr.name, KMOD_MAX_NAME)) {