+++ Lilu/Sources/kern_start.cpp
@@ -55,7 +55,7 @@ void PRODUCT_NAME::stop(IOService *provider) {
 
 Configuration ADDPR(config);
 
-bool Configuration::performInit() {
+bool Configuration::performEarlyInit() {
 	kernelPatcher.init();
 
 	if (kernelPatcher.getError() != KernelPatcher::Error::NoError) {
@@ -65,6 +65,31 @@ bool Configuration::performInit() {
 		return false;
 	}
 
+	KernelPatcher::RouteRequest request {"_PE_initialize_console", initConsole, orgInitConsole};
+	if (!kernelPatcher.routeMultiple(KernelPatcher::KernelID, &request, 1, 0, 0, true, false)) {
+		SYSLOG("config", "failed to initialise through console routing");
+		kernelPatcher.deinit();
+		kernelPatcher.clearError();
+		return false;
+	}
+
+	return true;
+}
+
+int Configuration::initConsole(PE_Video *info, int op) {
+	DBGLOG("config", "PE_initialize_console %d", op);
+	if (op == kPEBaseAddressChange && !atomic_load_explicit(&ADDPR(config).initialised, memory_order_relaxed)) {
+		IOLockLock(ADDPR(config).policyLock);
+		if (!atomic_load_explicit(&ADDPR(config).initialised, memory_order_relaxed)) {
+			DBGLOG("config", "PE_initialize_console %d performing init", op);
+			ADDPR(config).performCommonInit();
+		}
+		IOLockUnlock(ADDPR(config).policyLock);
+	}
+	return FunctionCast(initConsole, ADDPR(config).orgInitConsole)(info, op);
+}
+
+bool Configuration::performCommonInit() {
 	lilu.processPatcherLoadCallbacks(kernelPatcher);
 
 	bool ok = userPatcher.init(kernelPatcher, preferSlowMode);
@@ -86,6 +111,19 @@ bool Configuration::performInit() {
 	return true;
 }
 
+bool Configuration::performInit() {
+	kernelPatcher.init();
+
+	if (kernelPatcher.getError() != KernelPatcher::Error::NoError) {
+		DBGLOG("config", "failed to initialise kernel patcher");
+		kernelPatcher.deinit();
+		kernelPatcher.clearError();
+		return false;
+	}
+
+	return performCommonInit();
+}
+
 int Configuration::policyCheckRemount(kauth_cred_t, mount *, label *) {
 	ADDPR(config).policyInit("mac_mount_check_remount");
 	return 0;
@@ -94,6 +132,12 @@ int Configuration::policyCheckRemount(kauth_cred_t, mount *, label *) {
 int Configuration::policyCredCheckLabelUpdateExecve(kauth_cred_t, vnode_t, ...) {
 	ADDPR(config).policyInit("mac_cred_check_label_update_execve");
 	return 0;
+}
+
+void Configuration::policyInitBSD(mac_policy_conf *conf) {
+	DBGLOG("config", "init bsd policy on %u in %d", getKernelVersion(), ADDPR(config).installOrRecovery);
+	if (getKernelVersion() >= KernelVersion::BigSur)
+		ADDPR(config).policyInit("init bsd");
 }
 
 #ifdef DEBUG
@@ -170,7 +214,7 @@ bool Configuration::getBootArguments() {
 
 	betaForAll = checkKernelArgument(bootargBetaAll);
 	debugForAll = checkKernelArgument(bootargDebugAll);
-	isUserDisabled = checkKernelArgument(bootargUserOff);
+	isUserDisabled = checkKernelArgument(bootargUserOff) || getKernelVersion() >= KernelVersion::BigSur;
 
 	PE_parse_boot_argn(bootargDelay, &ADDPR(debugPrintDelay), sizeof(ADDPR(debugPrintDelay)));
 
@@ -256,6 +300,15 @@ bool Configuration::registerPolicy() {
 	if (policyLock == nullptr) {
 		SYSLOG("config", "failed to alloc policy lock");
 		return false;
+	}
+
+	if (getKernelVersion() >= KernelVersion::BigSur) {
+		if (performEarlyInit()) {
+			startSuccess = true;
+			return true;
+		} else {
+			SYSLOG("config", "failed to perform early init");
+		}
 	}
 
 	if (!policy.registerPolicy()) {